adk_inventory_agent_old/requirements.txt
# Core ADK Requirements
google-adk>=0.1.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Date/Time Handling
python-dateutil>=2.8.0

# Configuration Management
python-dotenv>=1.0.0

# Optional: Enhanced Analytics (uncomment if needed)
# scikit-learn>=1.3.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.15.0

# Development Tools (optional)
# jupyter>=1.0.0
# pytest>=7.0.0
